+++ NEWS.md
@@ -1,7 +1,11 @@
 ## Release 0.3
 __Breaking changes__
 
-* It is now ONLY possible to convert an `Option{T1}` to an `Option{T2}`, if `T1 <: T2`. This and the following change means type conversions now error in a type stable manner.
+* `Option{T}` has been revamped and is now an alias of `Result{T, Nothing}`. As a consequence, most code that used `Option` is now broken.
+* `Option`-specific functions `Thing`, `is_none`, `expect_none` and `unwrap_none` have been removed.
+* `none` is now a const for `Err(nothing)`, but works similarly.
+* `none(::Type{T})` now creates an `Err{T, Nothing}(nothing)`, e.g. `none(T)` behaves like `None{T}()` before.
+* `some(x)` creates an `Ok{typeof(x), Nothing}(x)`, e.g. it's like `Thing(x)` before.
 * It is now ONLY possible to convert a `Result{O1, E1}` to a `Result{O2, E2}` if `O1 <: O2` and `E1 <: E2`.
 * `@? x` when `x` is a `Result{O, E}` containing an `Err` with value `v` now evaluates to `return Err(v)` instead of `return x`. This allows a value of one `Result` type to be propagated to another.
 * The function `and_then`, now has the signature `and_then(f, ::Type{T}, x::Union{Option, Result})`, in order to prevent excessive type instability.
@@ -13,7 +17,6 @@ __New features__
 * New function `flatten`. Turns an `Option{Option{T}}` into an `Option{T}`.
 * New function: `unwrap_err(::Result)`. Unwraps the wrapped error type of a `Result`, and throws an error if the `Result` conains an `Ok`.
 * New function: `expect_err(::Result, ::AbstractString)`. Similar to `unwrap_err`, but throws an error with a custom message.
-* `None` is now exported.
 
 ## Release 0.2
 __Breaking changes__